+++ widget/android/CompositorWidgetParent.cpp
@@ -38,9 +38,16 @@ RefPtr<VsyncObserver> CompositorWidgetParent::GetVsyncObserver() const {
 }
 
 void CompositorWidgetParent::OnCompositorSurfaceChanged() {
-  mSurface = java::GeckoServiceGpuProcess::RemoteCompositorSurfaceManager::
-                 GetInstance()
-                     ->GetCompositorSurface(mWidgetId);
+  java::GeckoServiceGpuProcess::RemoteCompositorSurfaceManager::LocalRef
+      manager = java::GeckoServiceGpuProcess::RemoteCompositorSurfaceManager::
+          GetInstance();
+  java::sdk::SurfaceControl::LocalRef surfaceControl =
+      manager->GetCompositorSurfaceControl(mWidgetId);
+  if (surfaceControl) {
+    mSurface = java::sdk::Surface::FromSurfaceControl(surfaceControl);
+  } else {
+    mSurface = manager->GetCompositorSurface(mWidgetId);
+  }
 }
 
 }  // namespace widget